added lib folder and moved files around
